refactor(solver): isolate relation cache identity types (#14351)
